MB-52311: [8/n] Pause / Resume Bucket: EPBucket cancellation
Extend cancellation support into EPBucket's prepareForPause() method.

For phases of this method which we expect to be long-running, add
support for checking the cancellation status of the Pause request, and
if cancelled undo any work done for pausing the bucket; returning
'cancelled' and in the same state as originally (before pausing).

Change-Id: I8c6a263c6f0af498e91f97c805602540019790a1
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/181728
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>