MB-52311: [5/n] Pause / Resume Bucket: Make pause() blocking
Change the mcbp Pause() request to be blocking - i.e. it waits until
the bucket has completed pausing (or failed for some reason) before it
returns a status code to the client.

This avoids ordering issues seen when Pause() was non-blocking (and
returned "Success" before the background task to perform it had
actually been executed) - request sequences like "Pause, Pause,
Resume" could end up getting scheduled internally as "Pause, Resume,
Pause" due to Pause running on a background thread.

This also introduces a new status code - BucketPaused (0x50) which is
returned if one attempts to perform an operation on a Bucket which is
already paused - such as calling pause() again.

Now pause() is blocking the testcase which was disabled due to failing
intermittently (PauseResumeTest/PauseFailsWhenPaused) has been
re-enabled.

Change-Id: Iaee1e036f6256dce514f63ed6343bbc105254f04
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/181463
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>