MB-54161: Fix input buffer handling in test client impl
Don't try to reallocate the input buffer in the callback from
folly as that beaks the ownership model for buffer handling
in folly. Instead allocate a new buffer and tell folly to
use the new buffer.

To make sure that we don't mess around with the buffers at
the same time as folly we may just schedule another callback
to process the received data.

Change-Id: I0e96e0ddc96109e91bc85bb801744cbec163bd59
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/181447
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>