MB-54053: RangeScan don't update continueRunState after state change
RangeScan::continueRunState is a copy of the state required to continue
a scan that is used by the task (so it doesn't need to frequently
lock/unlock to evaluate limits etc...). This copy must only be used
whilst the state is continue.

This commit fixes an issue where setStateIdle was modifying
continueRunState after the RangeScan had been set to Idle and
as demonstrated by the MB this lead to a subsequent continue
overlapping with this code leading to an exception.

The fix is applied in setStateIdle where the code now copies the
cookie before changing state and never touches continueRunState
after the state change.

A unit test reproduced the issue after some restructuring which allowed
the overlapping code to be executed in a single threaded test.

Change-Id: Ia5fcb46d3f25c1573bdfed63875b58bc09affd9c
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/181460
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>