MB-35297: Stop and cancel a RangeScan when disconnected during a scan
Prior to this commit, if a client disconnects after a continue request
the RangeScanContinueTask has no mechanism to break the scan, it could
"spin" and progress the iterator for some time.

This commit adds a disconnected detection path from the
Cookie::sendResponse function which the scan loop is using to push
data to the connection. If a disconnection is detected the policy
is to cancel the RangeScan as the iterator may have been progressed
forwards and the client has no way to correctly resume the scan.

Change-Id: Id783a891a520100659c932954add83841d0d27f2
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/181469
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>