MB-52081: Rename kv_ep_kv_size{for=hashtable} into kv_ep_key_value_size
The stat accounts for all items metadata, keys and values allocated in
the system. All value blobs are accounted regardless of whether they
are referenced by HashTables or already ejected.

Note that the existing ep_ht_item_memory accounts exactly the size of
StoredValue objects + keys + blobs referenced by the HashTable, so we
have already a quantity that tracks the exact mem usage in HTs. Also,
we have the vb_{state} variants of the same.

Note
Before this patch prometheus exposes a
kv_memory_used_bytes{for="hashtable"}. That used to map the KV
kv_ep_kv_size_bytes{for="hashtable"}.
By this patch that is entirely removed in prometheus, which now exposes
the plain kv_ep_key_value_size_bytes.
cbstats is unchanged, that still exposes the old ep_kv_size.

Change-Id: I333510cedb3f3b2d6f67c1dd6b8a524cadd59b2b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/181402
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>