MB-46890: Filter out events in the front end thread
Previously we would unconditionally push all (externally
generated) audit events to the audit daemon, which would
then iterate over all of the events and drop disabled and
events which should be filtered out before writing the
remaining events to disk.

This patch change the logic to check if the event is
enabled and doesn't match the "ignore filter" before
submitting the audit event to the backend. By doing so we'll
reduce the likelihood of dropping events in the case where
the audit backed is getting behind on writing the audit trail
to disk.

Change-Id: I447e30e2b66cb9e383c5a03e3ea292e6c5aeee3a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/181671
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>