MB-54196: Cancel RangeScan in DeleteBucketTask at shutdown
Before this patch a RangeScan is cancelled by spwaning an instance of
RangeScanContinueTask and deferring the cancellation to it.

The problem with that approach is that the RangeScanContinueTask might
be halted at shutdown before it manages to run and do its work.

One way of preventing that from happening would be setting
completeBeforeShutdown=true for the task.
But we do something different here. Given that the entire
bucket-deletion EP-step is executed already in a bg-task (the
Core_DeleteBucketTask), then we clean up RangeScans inline there.
So no need to spawn any RangeScanContinueTask for that, cancellation
isn't blocking any frontend/worker thread.

New code already covered by the test_range_scan_no_cancel test
introduced at https://review.couchbase.org/c/kv_engine/+/181151
(MB-53953).

*Note*
Why the same test wasn't enough to cover the previous behaviour while
it's good as coverage now?
That's because the previous behaviour wasn't deterministic, as
RangeScanContinueTask(completeBeforeShutdown=false) means that the
task might or might not complete before the thread-pool is shutdown.

Change-Id: Id2687265baebcfd7760a715325d4427237c0e13b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/181788
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>