Add WeakPtrBag
An unordered collection which is implemented as a thread-safe vector of
weak_ptrs.

The motivation for this is to have a method like 'getAllExistingTasks',
such that we can orchestrate the execution of tasks relevant to unified
caching across all buckets without making the core aware of the
individual tasks' existence.

Change-Id: Id91a8153611909c300c2daeafea1fd8bcee6dec4
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/181794
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>