MB-46890: Add support for per-bucket audit filter
This patch adds (internal) support for allowing a per-bucket
configuration for audits.

The external API (configuration etc) remains unchanged, but
internally we wrap the old configuration into the new configuration
and initialize the data structures with the new configuration
format.

In the new configuration one may enable events on a per-bucket
level, so one may enable document access audit in bucket A and
disable it in bucket B.

Previously there was a single list of users to ignore audit
events, and that was used to filter _ALL_ different audit events.
In the new model one may specify a list of events where the
user should be ignored. Which means that you could avoid having
document read being audited for user X, but any document write
would be audited for user X.

The same model may be pushed down to scope and collections, but
that is not implemented yet.

Change-Id: Ie9124f3edb9968512ee7f8799196812556a7af71
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/181803
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>