Build the fuzz tests when we don't have the fuzzer
The fuzz tester works by creating a function with a given
signature which gets called from the fuzzer framework.

Our code does however contain a lot of unrelated logic
(initializing various components etc) which is unrelated
from the fuzzer framework. To make it easier to debug
the framework (and fix such problems) always build
the program and just call the fuzzer function (which
would cause the initialization phase to get called)

Change-Id: If2a913b100f1d2eeb5d990d4fd95834f9c398ce3
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/181941
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>