MB-54294: Add initial kv_logical_data_size_bytes metric
To inform scaling decisions for elixir, the control plane requires an
estimate of how much memory all data on disk would require if it were
resident.

The sum of the collection data size is currently _almost_ the desired
value; under magma it tracks the size of the value _as stored into KV_,
whether compressed or not.

For highly compressible datasets, this may overestimate how much memory
would be required for an active compression bucket to hold that data.

_For now_ exposing this sum to allow the control plane to begin
testing.

Once MB-53858 is complete, this may be replaced with a better
implementation.

Change-Id: Idffb48cb8b4403168cc67a2d06d1387013656327
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/181962
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>