Refactor: move auditd test to memcached_testapp
Checking for version and may be easily be tested together
with all of the other audit events as we have the infrastructure
of reading the audit files in place.

One may argue that the previous method was a bit more "unit test
like", but we had to mock a lot of classes and infrastructure
which made it easier to add those unit tests into the bigger
framework.

Change-Id: I7cb259b126cbc0d375ff83d03ce22fa9d10d4e2a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/182021
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>