MB-53304: Enforce holding of stateLock in VBucket::queueItem [8/n]
MB-53304 suggests we ensure that the stateLock is held by passing a ref
to it in queueItem (and the call hierarchy of queueItem).

This change ensures that we hold the VBucket state lock when creating a
pending SyncWrite.

In a future change I will pass this ref down to queueItem.

Change-Id: I8a31582b8dbf97642a55c32c1d9028df0bf99e6d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/182030
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>