MB-53304: Enforce holding of stateLock in VBucket::queueItem [9/n]
MB-53304 suggests we ensure that the stateLock is held by passing a ref
to it in queueItem (and the call hierarchy of queueItem).

This change updates the following VBucket methods to take a ref to the
state lock:
- set
- setWithMeta
- add
- replace
- deleteItem
- deleteWithMeta
- getAndUpdateTtl

In a future change I will pass this ref down to queueItem.

Change-Id: I60e984df9b0213d1909523f6964228f752e18b7d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/182028
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>