MB-46890: Generate the list of enabled audits
The list of the enabled audit events gets built up at runtime
by using the "compiled in" default values and allow to override
them in the "event_states" in the configuration.

Currently we try to generate the "new" audit filter by converting
the config into the new config and as part of the refactor the
"compiled in" enabled events got lost (and there was no unit
test to verify that). This patch adds that back (with a test)

Change-Id: I4c54951adac1aac953ee4bedca864f8dfabdf3b4
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/182098
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>