MB-54340: Convert builtin and couchbase to Local domain
Audit filtering only allows to configure users by specifying
"couchbase" or "external" as the domain in the filtering
configuration. The audit daemon maps "couchbase" to local
when it reads the configuration, but some modules still
provide "builtin" (as seen in MB-54340) and I wouldn't
be surprised if someone also use couchbase.

To allow these events to be submitted and filtered allow
"builtin" and "couchbase" to be used as aliases for
"local".

Change-Id: I82025df7fb06c644b1e544741dfaf48e9667b382
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/182121
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>