Validate audit events before dumping them
Previously we did not validate the user input (that it contained
the mandatory fields it said it would contain). This patch
add the validation and will log (and drop) the events missing
the mandatory fields when developer asserts is enabled so
that it is detected by the developers unit tests during
development so they may fix the audit they submit.

Change-Id: Ia932f36fff031b7ebb30f0b9f1acd1d658a2e7b4
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/182123
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>