Update domain description for audit
When adding support for filtering the UI only allows to specify
two different domains (couchbase and external). Update the
documentation for the audit events to match this.

Note that internally the "couchbase" domain is referred as
"local" (but the parser allows the aliases builtin and couchbase
for now). Ideally we would change all occurrences to use a
single name all over the product.

Change-Id: Id5321d9ea975a173ffd07f181a1f3b15aeed35d7
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/182266
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>