MB-53053: Allow buckets in the Ready state to be kept alive
Add a new `tryAssociateBucket` to the ServerBucketIface that will
allow us to safely interact with buckets outside of the bucket's own
context. Without a way to keep the bucket alive, there is no safe way
to schedule a task or get buckets stats for another bucket.

We want to be able to do this so that we can orchestrate ItemPager
across buckets sharing memory quota.

Change-Id: Iefd112ac40f2a6448c9a6ef02640276a27221f5f
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/182285
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>