Notify bucket cond variable for Bucket::State::Pausing
Notify the condition variable when the last client disconnects
when the bucket state is Pausing. This wakes the thread waiting
for the clients to disconnect immediately instead of waiting
for the timeout.

(The pause code wakes every second to notify clients and
cancel operations and log which clients we're waiting for)

Change-Id: I7b511a373473232e71adea18ac6fd3902bbc4f37
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/182316
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>