MB-54357: Skip select-bucket in 'cbstats tasks-all'
Recently we have introduced the 'tasks-all' group in cbstats. That
outputs the running tasks for all ready buckets, plus NoBucket.
https://review.couchbase.org/c/kv_engine/+/180238 for reference.

Given the nature of the command, there is no need to specify any
bucket in the call. But unfortunately the current behaviour in
cbstats is that 'default' is used if the user doesn't provide the
bucket name. The result is that the current cbcollect call like:

  ./cbstats host:port -u user -p password tasks-all

fails by

  No access to bucket:default - permission denied or bucket does not exist.

Notes:
  - An alternative fix would be removing the 'default' bucket behaviour
    in cbstats, but that's likely to break users out there.
  - ns_server's cbcollect doens't require any further change by this

Change-Id: I6788122f6130dc54ea031569ee443fd3406e5345
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/182313
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>