MB-53053: Add CrossBucketVisitorAdapter for cross-bucket visitation
The CrossBucketVisitorAdapter can be passed in a list (map) of
bucket -> visitor. The visitors are wrapped in an
SingleSteppingVisitorAdapter, which is a GlobalTask which visits one
VBucket and pauses. The cross-bucket adapter will then wake up each
individual bucket's visitors either in a sequential or round-robin
fashion until all visitors have completed.

The plan is to use this cross-bucket visitor adapter to run the
PagingVisitor tasks across multiple buckets in a fair way.

Change-Id: Ie7a4afb11ecadd8e77b9fe9142c69505a82d2587
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/182442
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>