MB-53304: Enforce holding of stateLock in VBucket::queueItem [10/n]
MB-53304 suggests we ensure that the stateLock is held by passing a ref
to it in queueItem (and the call hierarchy of queueItem).

This change makes fetchValidValue and its callers pass down the lock on
the vbucket state.

In a future change I will pass this ref down to queueItem.

Change-Id: I81d1d78a1758421f30a92f3b8c021d5d646bdec8
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/182436
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>