MB-53053: Make ItemPager inherit from NotifiableTask
Also had to add NotifiableTask::wasNotifiedManually.

The ItemPager task exposes no other behaviour other than that it can be
woken up and that wakeup being debounced.

By inheriting from NotifiableTask, we remove the duplicated debouncing
code and we can now refer to this task by a `NotifiableTask*`. This
will allow us to define a cross-bucket ItemPager task in the future
and refer to it by the same pointer type.

Change-Id: I8de705818c10b2aa5c2e7dd556fa3d08e13c6117
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/182555
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>