MB-53053: Add SingleSteppingVisitorAdapter for cross-bucket eviction
The planned approach for cross-bucket item eviction is to evict items
from VBuckets in a round-robin fashion.

This can be accomplished by having an instance of this adapter task
scheduled per bucket and by manually wake()-ing each bucket's task in
a round-robin fashion.

This adapter will allow us to evict items from a:1 (bucket:vb) callback
into some common synchronisation point, schedule b:1, callback again,
schedule a:2 and so on...

*Alternative approaches*

We could use a GlobalTask scheduled on some non-engine taskable
(NoBucketTaskable or some new cross-bucket paging taskable). If we were
to do that, we'd have to maintain the state of each bucket's
PagingVisitors (or create a new cross-bucket PagingVisitor) and would
have to ensure that we're not starving threads, handling engine
shutdown etc.

We could also just set the existing PagingVisitors running as we
currently do, for all buckets sharing quota, and stop when the
sum(mem_used) < sum(LWM). But that could mean unfair eviction in the
sense that large HTs with a similar number of evictable items at the
same MFU might be slower to evict from, thereby always penalising
small HTs.

Change-Id: Ic92099b27e1c00933d346f4a0688f7c5e8282133
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/182318
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>