MB-54516: Add Collection's history setting to VB::Manifest
Make the collection's history configuration (which only affects
deduplication) available via the VB::Manifest - this will later
be used to configure deduplication for writes to the vbucket.

Limitation of this patch:

There is no replication of the VB::Manifest value, only an active
vbucket will know the value and state change will not adjust the
value. These limitations will be addressed separately.

Change-Id: I9956afd3c40f171ac77f33dfa0a457ec98e006d3
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/182644
Well-Formed: Restriction Checker
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>