MB-54516: Fix VBucketManifestTest.duplicate_cid_different_scope
This test currently passes, but not for the reason it is covering.

The test expects a manifest update to fail, it does, but only because
no system-events were generated from the update. The test was
expecting an validation failure to due to a duplicate collection
(collection in different scopes).

Add the "shop1" scope before adding the collection and the test now
covers the intended validation path.

Change-Id: I81c30825a525d38507284ebeb950b0244fae7126
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/182767
Well-Formed: Restriction Checker
Reviewed-by: James H <james.harrison@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>