MB-54516: Update SystemEvent to include history configuration
Update the schema in events.fbs to include a new bool field for storing
the collections history setting.

DCP will now include the value a collection was created with and
provided the consumer/producer have negotiated FlatBuffers events the
replica is updated to read this value when processing a
CreateCollection.

Change-Id: I83f8649a7efea31abd9aaadc7f84854a192e972d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/183247
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>