MB-54587: Don't fail startup if we can't get timestamp
Previously the audit daemon would refuse to start if it encountered
any "problems" trying to read, parse or pick out the timestamp from
the old audit file.

This patch change the behavior to "ignore" all such errors and use
the "current time" in the filename instead of the oldest event in
the file when it encountered any problems.

It'll still fail if there is an IO Error reported from the OS which
won't go away after retrying

Change-Id: Ief78db2ecf1538aeff3ac13f548fe718fa2ed57c
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/182825
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>