MB-53053: Expose the NoBucketTaskable to use in the quota sharing pager
The quota sharing item pager will need a common synchronisation point.
It makes sense to use the NoBucketTaskable for that, as it is outlives
all buckets. A custom taskable defined in ep-engine would not get
unregistered before the ExecutorPool is shutdown (the executor pool is
created in daemon/memcached.cc).

This change exposes the NoBucketTaskable as an ExecutorPool's "default"
taskable.

Change-Id: I229b6949cce27334e5106aad0a31f5500c1d6ae6
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/183328
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>