MB-54516: DCP: Optionally produce SystemEvents with FlatBuffers values
Add a DCP Control option which changes the format of the value that is
sent with each system-event. Instead of the custom mcbp structures the
control enables FlatBuffers structures using the following schema:

  engines/ep/src/collections/events.fbs

With this commit, the KV DCP consumer now attempts to enable FlatBuffers
system events (requiring compatible producer).

The intention of this patch is to provide a cleaner route forward for
updating the metadata of a collection and to later permit modifications
to collection or scope metadata. The FlatBuffers code paths are less
burdensome to maintain compared to the custom mcbp structures added in
7.0.

The majority of the Collection DCP test coverage now uses the
FlatBuffers mode except for a small number of test that use more direct
calls into the DCP consumer, giving coverage of the mcbp defined
structures.

Change-Id: Ie2bb6e09a8d9bf67f2d1aec673b6aacf5b569c78
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/183246
Well-Formed: Restriction Checker
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>