MB-54386, MB-54455, MB-54452: Add ConnectionLimitMode::Recycle
Add a new mode for how the system should behave when one reach the
maximum number of user connections. This logic does _not_ apply to
the system interfaces (and the rest of the commit message only discuss
the user interfaces). The system continues to have a maximum number
of user connections which is a hard limit we'll never exceed.
Once the system hits the hard limit we'll immediately close the
incomming connection.

By default we continue to use the old mode of operation which is
called "disconnect".

The other mode of operation is "recycle"; and once enabled the
system will start disconnecting clients (picking the victim on
a per-thread LRU) once we hit 99% of the user connections.
Closing connections may involve "some work" so there is a
possibility that the incomming rate exceeds the rate closing
the connections. If we end up with 100% of the user connections
being used the system will disconnect the new connections
immediately just like in "disconnect" mode (note that the system
is still trying to close connections and once 1 connection gets
closed the next connect attempt will use the newly freed slot etc)

To enable the new mode the following entry needs to be inserted in
memcached.json:

    "connection_limit_mode" : "recycle"

If the entry is absent (or set to "disconnect") the system will use
the good old mode we've learned to love over the years.

It is hard to predict a number which would be correct in all cases
so it is possible to provide an absolute number for the number
of connections to allow instead of 1%. To do so insert the
following entry in memcached.json:

    "free_connection_pool_size" : <number>

Change-Id: I477ed3ab60b3c0cdf92f59d915f1c62e070cffa9
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/182979
Well-Formed: Restriction Checker
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>